fix:database issues factory_bot rspec tested
