--- app/controllers/products_controller.rb
@@ -1,38 +1,0 @@
-class ProductsController < ApplicationController
-  def index
-   # @products = Product.all
-   render json: Product.all
-  end
-
-  def new
-    @product = Product.new
-  end
-
-  def create
-    @product = Product.create(product_params)
-  # @product.save
-
-   # redirect_to @product
-  # @products = Product.all
-  end
-
-  def show
-    @product = Product.find(params[:id])
-    render json: @product
-
-  end
-  def update
-    @product.update(product_params)
-   # redirect_to @product
-  end
-
-  def edit
-    @product = Product.find(params[:id])
-  end
-
-  private
-
-  def product_params
-    params.require(:product).permit(:title, :short_description, :price)
-  end
-end
