+++ config/application.rb
@@ -9,8 +9,8 @@ Bundler.require(*Rails.groups)
 module BetaShop
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    
-    config.autoload_paths += %W(#{config.root}/lib)
+
+    config.autoload_paths += %W[#{config.root}/lib]
     config.load_defaults 7.0
 
     # Configuration for the application, engines, and railties goes here.
@@ -20,6 +20,5 @@ module BetaShop
     #
     # config.time_zone = "Central Time (US & Canada)"
     # config.eager_load_paths << Rails.root.join("extras")
-    
   end
 end
