+++ config/database.yml
@@ -11,15 +11,15 @@ default: &default
 
 development:
   <<: *default
-  username: <%= Rails.application.credentials.database[:username] %>
-  password: <%= Rails.application.credentials.database[:password] %>
+  username: istiak
+  password: istiak
   host: localhost
   database: beta_development
 
 test:
   <<: *default
-  username: <%= Rails.application.credentials.database[:username] %>
-  password: <%= Rails.application.credentials.database[:password] %>
+  username: istiak
+  password: istiak 
   host: localhost
   database: beta_test
