+++ config/routes.rb
@@ -5,5 +5,10 @@ Rails.application.routes.draw do
 
   # Defines the root path route ("/")
   # root "articles#index"
-  resources :products
+
+  namespace :api do
+    namespace :v1 do
+      resources :products
+    end
+  end
 end
