+++ spec/requests/products/create_spec.rb
@@ -6,7 +6,6 @@ describe 'Creates Product' do
     {
       product:
       {
-
         title: 'Create a HTML Page',
         short_description: 'This should be responsive',
         price: 200
@@ -16,9 +15,12 @@ describe 'Creates Product' do
 
   context 'With valid params' do
     it 'creates a Product ' do
-      post products_path, params: valid_params
+      # binding.pry
+      post api_v1_products_path, params: valid_params
+      # binding.pry
       expect(Product.count).to eq(1)
-      #expect(Product.list.count).to eq(1)
+      # binding.pry
+      # expect(Product.list.count).to eq(1)
       expect(status).to eq(204)
     end
   end
